MMFMWK-9597 Fix the UndefinedBehaviorSanitizer issue.
Andorid UndefinedBehaviorSanitizer reports:
MP4Parser.c:4204:20: runtime error: unsigned integer overflow:
MP4Parser.c:4434:37: runtime error: unsigned integer overflow:
MP4OrdinaryTrackReader.c:166:27: runtime error: unsigned integer overflow:
MP4Parser.c:3092:22: runtime error: unsigned integer overflow:

+++ src/mp4_parser/lib/src/MP4OrdinaryTrackReader.c
@@ -163,7 +163,7 @@ static MP4Err getNextAccessUnit(struct MP4TrackReaderStruct* self, MP4Handle out
     *outDTS = sampleDTS + self->segmentBeginTime;
 
     if (sampleFlags & MP4MediaSampleHasCTSOffset)
-        *outCTS = *outDTS + sampleCTSOffset;
+        *outCTS = (u64)((s64)(*outDTS) + sampleCTSOffset);
     else
         *outCTS = *outDTS;
